backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
openai==0.28.1
langchain==0.0.350
chromadb==0.4.18
sentence-transformers==2.2.2
pytest==7.4.3
httpx==0.25.2
PyPDF2==3.0.1 